+++ src/shared/SideMenu/components/SideMenuItem.tsx
@@ -1,4 +1,4 @@
-import { useRouter } from 'next/dist/client/router';
+import { useRouter } from 'next/router';
 
 interface SideMenuItemProps {
   title: string;
@@ -16,7 +16,7 @@ export default function SideMenuItem({ title, link }: SideMenuItemProps) {
   return (
     <div className='w-full h-74 px-20 py-24 cursor-pointer transition-colors duration-300 ease-in-out hover:bg-gray-100 active:bg-gray-200'>
       <p
-        className='font-[Pretendard] text-16 font-semibold leading-26'
+        className='font-primary text-16 font-semibold leading-26'
         onClick={handleItemClick}
       >
         {title}
